backend/requirements.txt
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
easydict==1.13
flatbuffers==25.2.10
fonttools==4.56.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
inflection==0.5.1
insightface==0.7.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lazy_loader==0.4
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
onnx==1.17.0
onnxruntime==1.21.0
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
prettytable==3.16.0
protobuf==6.30.2
psycopg==3.2.6
pydantic==2.11.1
pydantic_core==2.33.0
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
simsimd==6.2.1
six==1.17.0
sqlparse==0.5.3
stringzilla==3.12.3
sympy==1.13.3
threadpoolctl==3.6.0
tifffile==2025.3.13
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13